--HG--
branch : yt

+++ yt/analysis_modules/cosmological_observation/light_cone/light_cone.py
@@ -298,7 +298,7 @@ class LightCone(CosmologySplice):
                              "or a tuple of type (float, str).")
         
         # Calculate number of pixels on a side.
-        pixels = (field_of_view / image_resolution).in_units("")
+        pixels = int((field_of_view / image_resolution).in_units(""))
 
         # Clear projection stack.
         projection_stack = []
